no faking int to float

+++ crates/nu-command/src/random/float.rs
@@ -90,7 +90,7 @@ fn float(
             let value = match range.end() {
                 Bound::Included(end) => random_range(range.start()..=end),
                 Bound::Excluded(end) => random_range(range.start()..end),
-                Bound::Unbounded => random_range(range.start()..i64::MAX as f64),
+                Bound::Unbounded => random_range(range.start()..f64::MAX),
             };
 
             Ok(PipelineData::Value(Value::float(value, span), None))
